refactor out styles
